LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/getBoundingClientRect-shy-expected.txt

123456­789012
123456­789012­345678­901234
12­456­789012
123­456­789012

PASS Range 0-5 should not include the hyphen.
PASS Range 1-5 should not include the hyphen.
FAIL Range 0-6 should not include the hyphen. assert_equals: expected 60 but got 55
FAIL Range 1-6 should not include the hyphen. assert_equals: expected 50 but got 45
PASS Range 0-7 should include the hyphen.
PASS Range 1-7 should include the hyphen.
PASS Range 0-8 should include the hyphen.
PASS Range 0-8 should include only the first hyphen.
FAIL Collapsed soft-hyphen should be 0 width. assert_equals: expected 0 but got 5
FAIL Rendered soft-hyphen should have a width. assert_equals: expected 10 but got 15
PASS Collapsed soft-hyphen in a span should be 0 width.
PASS Rendered soft-hyphen in a span should have a width.
